Added high level overview to 'getting started'
@justinmcgrath suggested that for new users a high-level overview of the platform would be helpful before diving in. More specific than the introduction to PEcAn, but digestible to a broad audience.

He pointed out that many software documentation sites have minimal code snippets. 

He also offered to take a whack at revising this further, so I am submitting as a pull request on a new branch pending input from him and others. I am not sure if this is the right place for this text but it should a) be easy to find and b) be available before a new user begins 'getting started'.